splitted models and repos into their own folders
