+++ IdentityServiceApi/Dal/Roles/RoleDal.cs
@@ -1,8 +1,9 @@
 ﻿using Core.Dal.Base;
+using Dal.Rights;
 
-namespace Dal.Models
+namespace Dal.Roles
 {
-    public record RoleDal: BaseEntity
+    public record RoleDal : BaseEntity
     {
         public string Name { get; init; } = null!;
         public List<RightDal> Rights { get; set; } = null!;
